initial port of run to libmamba and adding shell command to enter activated shell
