+++ micromamba/src/completer.cpp
@@ -1,7 +1,7 @@
 #include "mamba/api/configuration.hpp"
 
 #include "mamba/core/output.hpp"
-#include "run.hpp"
+#include "mamba/core/run.hpp"
 
 #include <CLI/CLI.hpp>
